requirements.txt
python-telegram-bot>=20.0,<22.0
fastapi==0.95.2
uvicorn[standard]==0.22.0
